Reorganize store reading into one file
