+++ test/diffAgainstStore.ts
@@ -2,7 +2,7 @@ import { assert } from 'chai';
 
 import {
   diffQueryAgainstStore,
-} from '../src/data/diffAgainstStore';
+} from '../src/data/readFromStore';
 
 import { writeQueryToStore } from '../src/data/writeToStore';
 
@@ -108,7 +108,9 @@ describe('diffing queries against the store', () => {
         store,
         query: unionQuery,
         variables: null,
-        throwOnMissingField: true,
+        options: {
+          returnPartialData: false,
+        },
       });
     }, /No fragment/);
   });
@@ -149,7 +151,9 @@ describe('diffing queries against the store', () => {
         store,
         query: unionQuery,
         variables: null,
-        throwOnMissingField: true,
+        options: {
+          returnPartialData: false,
+        },
       });
     });
   });
@@ -190,7 +194,9 @@ describe('diffing queries against the store', () => {
         store,
         query: unionQuery,
         variables: null,
-        throwOnMissingField: true,
+        options: {
+          returnPartialData: false,
+        },
       });
     });
   });
@@ -232,7 +238,9 @@ describe('diffing queries against the store', () => {
         store,
         query: unionQuery,
         variables: null,
-        throwOnMissingField: true,
+        options: {
+          returnPartialData: false,
+        },
       });
     });
   });
@@ -275,7 +283,9 @@ describe('diffing queries against the store', () => {
         store,
         query: unionQuery,
         variables: null,
-        throwOnMissingField: true,
+        options: {
+          returnPartialData: false,
+        },
       });
     });
   });
@@ -378,7 +388,9 @@ describe('diffing queries against the store', () => {
         store,
         query: simpleQuery,
         variables: null,
-        throwOnMissingField: true,
+        options: {
+          returnPartialData: false,
+        },
       });
     });
   });
